core: remove I/O from flaky DNS test
Moved the ResourceResolverFactory creation to DnsNameResolverProvider.
- logging class of the Factory is changed.

Resolves #4998
